This update allows SPINE applications to input the SPINE framework the baseStation port and the used mote platform (currently: "tinyos", "sf" and "emule" - see spine.SPINESupportedPlatforms) by specifying an application properties file in the new SPINEManager.getInstance method that therefore now becomes MANDATORY. This application properties file can be located anywhere and named arbitrarily, but MUST define at least the following properties: - "MOTECOM" (e.g. "serial@COM1:telosb", "sf@192.0.0.1:9002") - "PLATFORM" (currently: "tinyos", "sf" or "emule")
Moreover, by defining the further "SPINE_HOME" property, it's NOT anymore necessary to copy the 'defaults.properties' file in the application HOME folder.
This "SPINE_HOME" property has effect only if the application references the SPINE framework source project. 
Meanwhile, if the application references the SPINE.JAR library, the 'defaults.properties' is automatically and transparently loaded from the jar file.
NOTE: the SPINE.JAR is NOT supposed to be COMPRESSED.

Finally, applications are not anymore required to set the System property LOCALNODEADAPTER_CLASSNAME.

!!! IMPORTANT: ALL THE SPINE DEVELOPERS ARE STRONGLY ENCOURAGED TO UPDATE THEIR APPLICATION CODE BY AVOIDING USE OF DEPRECATED METHODS !!!
!!! See the SPINETest application for this refactoring update !!!